Improved restart ability for demo

+++ 020_advanced/140_multi_arch/clean.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
 export DOCKER_CLI_EXPERIMENTAL=enabled
-docker buildx use default
-docker buildx rm mybuilder
+if docker buildx inspect mybuilder; then
+    docker buildx use default
+    docker buildx rm mybuilder
+fi
